ADD AWS OAUTH AND AWS ACCOUNT LEVEL AUTHENTICATION
Private:General.ps1
Added enum called ApiTypes

Public:General.ps1:Set-DatabricksEnvironment:
Added parameters for use with AWS Oauth. Specifically, the Account ID, Client ID and Client Secret. All of which are needed for authentication.
Added a new parameter set to track this kind of login
This also sets the default Api type to WORKSPACE. This is primarily to provide error handling and messaging throughout the rest of the script(s) to intelligently deal with variations in the API's. This makes the whole think basically "modal" which I'm not a fan of, but I think this is a stepping stone, not a definitiive solution.
The  WORKSPACE and ACCOUNT API's have different requirements. The workspace API's are easier to construct, while the account API's need the Account ID supplied in order to function. For this reason the Account ID is tracked in the script global variables. I also place the Client ID and Client Secret into globals so that if (and when) our token expires, that the system can recognize this and simply request a new one. Not sure if this is a good idea.

Public:General.ps1:Invoke-DatabricksApiRequest
Added some logic to handle when a token expires. The code automatically gets  a new token if this happens. Again, not sure if this is the best idea.

GroupsAPI.ps1:Get-DatabricksGroup
Added logic to test if we're logged in at the workspace or account level and basically throw an error if we're logged in with the account API. Continues processing if we are logged in with workspace API.